fsperf: change the config format a little bit
Instead of having the device and directory in the mkfs and mount
commands, specify the device per area and then just the naked mkfs and
mount commands, and then we'll handle adding the device and mount point
as appropriate.

Also add the ability to set an ioscheduler for the device.

Signed-off-by: Josef Bacik <josef@toxicpanda.com>