fsperf: filter based on test class name
Disabled tests are filtered according to the actual python class name,
whereas the name of the test when used in the positional argument is
based on the value of the "name" attribute of each respective test's
class. Unify the logic so that positional argument name is also the
test's class name.

Signed-off-by: Nikolay Borisov <nborisov@suse.com>