fsperf: make coloring symmetrical for good/bad
Currently, it comes up red when the means differ by 2 stddevs in the
wrong direction, and green otherwise. I feel this unfairly biases the
insignificant greens as wins. Color wins and losses symmetrically, and
leave a neutral color for <2 stddev diffs.

Note that I still had to color it for neutral because texttable has a
bug where alignment breaks for colored text due to the escape code, so
it looked silly and unaligned without coloring the default as well.

Signed-off-by: Boris Burkov <boris@bur.io>