add fragmentation tests
Changes to the block group and extent allocator can result in
regressions or improvements to how much we fragment a filesystem under
certain workloads. We are concerned about data block groups which have
large amounts of free space, but can't support large allocations. This
is reflected in the metric "fragmentation percentage" which is
calculated per block_group as:

100 * (1 - (max_free_extent / total_free))

So if all the free space is one free extent, the frag pct is 0%, and if
it's the worst case (I think...?) of 4K / 512MB, it is 99.99%.

The new fragmentation tests are their own suite in frag_tests/ and can
be added to a run with -F.

Analysis is done by:
1. running btrd to dump the block groups/extents
2. parsing that output with a rust program that spits out stats (and can
   spit out Hilbert Curve fragmentation images as well, but fsperf
   doesn't have a front-end for that, see the code)

This is all integrated into fsperf and the table dumping logic