add on-demand compare; refactor testonly
Currently, to get the awesome text comparisons, we need to run a
testonly run. For constant regression testing and re-baselining, this
can be wasteful, and the tests can take quite a long time.

Instead, a comparison utility which compares two run purposes, so we can
compare any two runs by purpose (if the user chooses to use different
purposes). For the testonly mode, add a special purpose which also
greatly improves the testonly cleaning, as we can drop TESTONLY purpose
runs before and after a testonly run, ensuring it never dirties the real
purpose's data. Incidentally, this fixes a bug in compare_config
comparisons which didn't correctly filter by purpose.