measure performance of a cycle mount after tests
at the end of a test, it is possible our FS will be in some gnarly
state: fragmented, lots of BGs, lots of extents, lots of snapshots,
orphans, whatever. It is useful to make sure we don't regress umount and
mount in such states, so toss in a final cycle mount that we can time
and track.