less printing noise
the test enabled/disabled stuff is useful for debugging but now that the
stuff is more stable, it's really noisy. We don't have "debug logs" in
this tool and I don't think that's worth it either, so just kill the
noise. Easy to add back if we need to debug again in the future.

Also don't print empty tables for tests that lack data.