sleep after umount to allow background cleanup to complete
If a mount is the last user of an implicit loop device (e.g. created by
mkfs.btrfs <file> then mounting) destroying the loop device is
asynchronous. This screws us up if we try to mount again right after
before it's done, so we need to sleep (there's not a great way to wait
for it more explicitly, short of figuring out the loop device name and
udev games)

If the mount is explicitly on a loop device, then we don't need to wait,
since the owner of the loop device is responsible for cleaning it up
after the umount.

Therefore, it's sufficient to just check if our device is a block device
with S_ISBLK