show regression when baseline is zero
Instead of skipping all keys where the mean is zero (generally fine for
fio jobs that don't do one of read/write and always have zero), only
skip the ones where both the baseline and the experiment are zero.
Otherwise, it is a regression/improvement. In practice this really shows
up in the fragmentation metrics, but theoretically could show up in a
latency trace or even fio metric.